chore: reuse Collect module components
